+++ wondrous-app/components/Common/BountyBoard/index.tsx
@@ -22,7 +22,7 @@ import TaskCardMenu from 'components/Common/TaskCardMenu';
 import TaskCardPrivacy from 'components/Common/TaskCardPrivacy';
 import TaskCardStatus from 'components/Common/TaskCardStatus';
 import TaskPriority from 'components/Common/TaskPriority';
-import ActionModals from 'components/Common/TaskViewModal/ActionModals';
+import ActionModals from 'components/Common/TaskViewModal/ActionModalsComponent';
 import { hasGR15DEIIntiative } from 'components/Common/TaskViewModal/utils';
 import PodIconName from 'components/Common/PodIconName';
 import { CreateEntity } from 'components/CreateEntity';
