+++ wondrous-app/components/Common/MilestoneBoard/index.tsx
@@ -19,7 +19,7 @@ import TaskCardMenu from 'components/Common/TaskCardMenu';
 import TaskCardPrivacy from 'components/Common/TaskCardPrivacy';
 import TaskCardStatus from 'components/Common/TaskCardStatus';
 import TaskPriority from 'components/Common/TaskPriority';
-import ActionModals from 'components/Common/TaskViewModal/ActionModals';
+import ActionModals from 'components/Common/TaskViewModal/ActionModalsComponent';
 import { CreateEntity } from 'components/CreateEntity';
 import EmptyStateBoards from 'components/EmptyStateBoards';
 import CommentsIcon from 'components/Icons/comments';
